Removed audio to transcript
