--- app/__init__.py
@@ -1,10 +1,0 @@
-from flask import Flask
-
-def create_app():
-    app = Flask(__name__)
-
-    # Register the Blueprint
-    from .routes import main
-    app.register_blueprint(main)
-
-    return app
